"small fixes for frontend"

+++ Bobs-Racing/Services/RaceService.cs
@@ -49,7 +49,7 @@ namespace Bobs_Racing.Services
                 {
                     // uses random to generate a random speed in a range based on min and max speed
                     // adding +1 to ensure the max value is included.
-                    checkpointSpeeds.Add(random.Next(athlete.LowestTime, athlete.FastestTime + 1));
+                    checkpointSpeeds.Add(random.Next((int)athlete.LowestTime, (int)(athlete.FastestTime + 1)));
                 }
                 // finds the sum of all checkpoint speeds to represent animals overall speed for a race
                 int totalSpeed = checkpointSpeeds.Sum();
